+++ fltk-sys/src/menu.rs
@@ -103,6 +103,13 @@ extern "C" {
     pub fn Fl_Widget_set_color(arg1: *mut Fl_Widget, color: libc::c_uint);
 }
 extern "C" {
+    pub fn Fl_Widget_measure_label(
+        arg1: *const Fl_Widget,
+        arg2: *mut libc::c_int,
+        arg3: *mut libc::c_int,
+    );
+}
+extern "C" {
     pub fn Fl_Widget_label_color(arg1: *mut Fl_Widget) -> libc::c_uint;
 }
 extern "C" {
@@ -364,6 +371,13 @@ extern "C" {
 }
 extern "C" {
     pub fn Fl_Menu_Bar_set_color(arg1: *mut Fl_Menu_Bar, color: libc::c_uint);
+}
+extern "C" {
+    pub fn Fl_Menu_Bar_measure_label(
+        arg1: *const Fl_Menu_Bar,
+        arg2: *mut libc::c_int,
+        arg3: *mut libc::c_int,
+    );
 }
 extern "C" {
     pub fn Fl_Menu_Bar_label_color(arg1: *mut Fl_Menu_Bar) -> libc::c_uint;
@@ -719,6 +733,13 @@ extern "C" {
 }
 extern "C" {
     pub fn Fl_Menu_Button_set_color(arg1: *mut Fl_Menu_Button, color: libc::c_uint);
+}
+extern "C" {
+    pub fn Fl_Menu_Button_measure_label(
+        arg1: *const Fl_Menu_Button,
+        arg2: *mut libc::c_int,
+        arg3: *mut libc::c_int,
+    );
 }
 extern "C" {
     pub fn Fl_Menu_Button_label_color(arg1: *mut Fl_Menu_Button) -> libc::c_uint;
@@ -1088,6 +1109,13 @@ extern "C" {
     pub fn Fl_Choice_set_color(arg1: *mut Fl_Choice, color: libc::c_uint);
 }
 extern "C" {
+    pub fn Fl_Choice_measure_label(
+        arg1: *const Fl_Choice,
+        arg2: *mut libc::c_int,
+        arg3: *mut libc::c_int,
+    );
+}
+extern "C" {
     pub fn Fl_Choice_label_color(arg1: *mut Fl_Choice) -> libc::c_uint;
 }
 extern "C" {
@@ -1433,6 +1461,13 @@ extern "C" {
 }
 extern "C" {
     pub fn Fl_Sys_Menu_Bar_set_color(arg1: *mut Fl_Sys_Menu_Bar, color: libc::c_uint);
+}
+extern "C" {
+    pub fn Fl_Sys_Menu_Bar_measure_label(
+        arg1: *const Fl_Sys_Menu_Bar,
+        arg2: *mut libc::c_int,
+        arg3: *mut libc::c_int,
+    );
 }
 extern "C" {
     pub fn Fl_Sys_Menu_Bar_label_color(arg1: *mut Fl_Sys_Menu_Bar) -> libc::c_uint;