+++ fltk-sys/src/misc.rs
@@ -103,6 +103,13 @@ extern "C" {
     pub fn Fl_Widget_set_color(arg1: *mut Fl_Widget, color: libc::c_uint);
 }
 extern "C" {
+    pub fn Fl_Widget_measure_label(
+        arg1: *const Fl_Widget,
+        arg2: *mut libc::c_int,
+        arg3: *mut libc::c_int,
+    );
+}
+extern "C" {
     pub fn Fl_Widget_label_color(arg1: *mut Fl_Widget) -> libc::c_uint;
 }
 extern "C" {
@@ -359,6 +366,13 @@ extern "C" {
 }
 extern "C" {
     pub fn Fl_Spinner_set_color(arg1: *mut Fl_Spinner, color: libc::c_uint);
+}
+extern "C" {
+    pub fn Fl_Spinner_measure_label(
+        arg1: *const Fl_Spinner,
+        arg2: *mut libc::c_int,
+        arg3: *mut libc::c_int,
+    );
 }
 extern "C" {
     pub fn Fl_Spinner_label_color(arg1: *mut Fl_Spinner) -> libc::c_uint;
@@ -668,6 +682,13 @@ extern "C" {
     pub fn Fl_Clock_set_color(arg1: *mut Fl_Clock, color: libc::c_uint);
 }
 extern "C" {
+    pub fn Fl_Clock_measure_label(
+        arg1: *const Fl_Clock,
+        arg2: *mut libc::c_int,
+        arg3: *mut libc::c_int,
+    );
+}
+extern "C" {
     pub fn Fl_Clock_label_color(arg1: *mut Fl_Clock) -> libc::c_uint;
 }
 extern "C" {
@@ -924,6 +945,13 @@ extern "C" {
 }
 extern "C" {
     pub fn Fl_Chart_set_color(arg1: *mut Fl_Chart, color: libc::c_uint);
+}
+extern "C" {
+    pub fn Fl_Chart_measure_label(
+        arg1: *const Fl_Chart,
+        arg2: *mut libc::c_int,
+        arg3: *mut libc::c_int,
+    );
 }
 extern "C" {
     pub fn Fl_Chart_label_color(arg1: *mut Fl_Chart) -> libc::c_uint;
@@ -1250,6 +1278,13 @@ extern "C" {
 }
 extern "C" {
     pub fn Fl_Progress_set_color(arg1: *mut Fl_Progress, color: libc::c_uint);
+}
+extern "C" {
+    pub fn Fl_Progress_measure_label(
+        arg1: *const Fl_Progress,
+        arg2: *mut libc::c_int,
+        arg3: *mut libc::c_int,
+    );
 }
 extern "C" {
     pub fn Fl_Progress_label_color(arg1: *mut Fl_Progress) -> libc::c_uint;