+++ fltk-sys/src/text.rs
@@ -103,6 +103,13 @@ extern "C" {
     pub fn Fl_Widget_set_color(arg1: *mut Fl_Widget, color: libc::c_uint);
 }
 extern "C" {
+    pub fn Fl_Widget_measure_label(
+        arg1: *const Fl_Widget,
+        arg2: *mut libc::c_int,
+        arg3: *mut libc::c_int,
+    );
+}
+extern "C" {
     pub fn Fl_Widget_label_color(arg1: *mut Fl_Widget) -> libc::c_uint;
 }
 extern "C" {
@@ -539,6 +546,13 @@ extern "C" {
 }
 extern "C" {
     pub fn Fl_Text_Display_set_color(arg1: *mut Fl_Text_Display, color: libc::c_uint);
+}
+extern "C" {
+    pub fn Fl_Text_Display_measure_label(
+        arg1: *const Fl_Text_Display,
+        arg2: *mut libc::c_int,
+        arg3: *mut libc::c_int,
+    );
 }
 extern "C" {
     pub fn Fl_Text_Display_label_color(arg1: *mut Fl_Text_Display) -> libc::c_uint;
@@ -1034,6 +1048,13 @@ extern "C" {
 }
 extern "C" {
     pub fn Fl_Text_Editor_set_color(arg1: *mut Fl_Text_Editor, color: libc::c_uint);
+}
+extern "C" {
+    pub fn Fl_Text_Editor_measure_label(
+        arg1: *const Fl_Text_Editor,
+        arg2: *mut libc::c_int,
+        arg3: *mut libc::c_int,
+    );
 }
 extern "C" {
     pub fn Fl_Text_Editor_label_color(arg1: *mut Fl_Text_Editor) -> libc::c_uint;
@@ -1611,6 +1632,13 @@ extern "C" {
 }
 extern "C" {
     pub fn Fl_Simple_Terminal_set_color(arg1: *mut Fl_Simple_Terminal, color: libc::c_uint);
+}
+extern "C" {
+    pub fn Fl_Simple_Terminal_measure_label(
+        arg1: *const Fl_Simple_Terminal,
+        arg2: *mut libc::c_int,
+        arg3: *mut libc::c_int,
+    );
 }
 extern "C" {
     pub fn Fl_Simple_Terminal_label_color(arg1: *mut Fl_Simple_Terminal) -> libc::c_uint;