+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Patch fltk to fix behavior of Fl_Win32_At_Exit() with mingw toolchain.
 - Fix android build which doesn't support Fl_Printer.
 - Fix WidgetExt::with_size() resizability.
+- Add WidgetExt::measure_label() and draw::measure().
 
 ## [0.10.4] - 2020-11-06
 ### Changes